fix: publishmap function
